Pass content to errorCatcher
The problem I was seeing was that apache was returning an html page on
some errors and there was no way to actually get the body of the
response in errorCatcher since only the error ("Response is not valid
JSON") was available. I noticed there was a content variable already
so am passing it to errorCatcher.